fix: target_compatible_with excluding arm64
The 'ci: update ci and exclude windows' commit added
'target_compatible_with' arguments to some of the tests restricting them
to linux but it also forced amd64 instead of allowing the tests to run
in both linux/amd64 and linux/arm64.

Before this fix, I would get SKIPPED tests when testing on linux/arm64
and now all the tests run.