fix: e2e MODULE.bazel.lock conflicts and repo name in copy.sh script (#87)
* fix: MODULE.bazel.lock conflicts

See https://github.com/bazelbuild/bazel/issues/20369

While working on refactoring, I kept hitting rebase conflicts due to
issues with the MODULE lock. I guess it's because the Bazel version in
e2e tests is much lower than the current one with the fix (7.2) but
still, I don't think it adds much to have the lock in e2e testing.

* fix: repo name in copy.sh script

PR #73 added the `_resolve` and this breaks the buildozer fix / autofix

* fix: remove dead locks

It seemed like 75afff9 in #47 added the new locks but as new files, that
is, the old ones were left behind.