refactor: remove is_satisfied_by
Once the validation of 'Provides' version is removed, we can simply use
version_lib.compare like we do with the non-virtual packages.

Note that the would-be arguments to is_satisfied_by are now inverted so
that the version comparison matches the order of the other version
comparisons in _resolve_package.