fix: copy examples/debian_snapshot to e2e/smoke test
These two have been the same but there's been changes to the examples
files that weren't replicated in the e2e test.

Copying examples/debian_snapshot to e2e/smoke and making a comment to
remark that it's important to keep these in-sync because testing in e2e
with at least the same base test helps in catching bugs like the
previous commit, where platforms repo should have been marked as
non-dev.