fix: #120 (and #115?) in MacOS
Bazel was throwing an error when running the tests in MacOS even when
the Linux tests were guarded with a target_compatible_with that was
restricted to Linux.

As far as I could tell, this was because running the bazel test with
//... was recursively evaluating other targets like
//examples/ubuntu_snapshot:_noble_index_json from oci_image and
oci_load. When I guard those rules with a target_compatible_with
everything works.

See https://github.com/GoogleContainerTools/rules_distroless/pull/136
for more context and information.

Also, seeing that e5f7dc0 also manually commented a test to fix CI
for macos so I guess #120 didn't break this, it was probably happening
since #115 added the new convenience targets.

Finally, maybe this is something that should be fixed in rules_oci, the
"inner targets" that it creates should probably be restricted to only
run in the `os` and `architecture` given for the image :-?