feat: support multiple URLs in sources (#134)
Fixes #113

The Debian snapshot mirror fails quite often and, while the Cloudflare
mirror has had some issues in the past (e.g. it was lagging replication
for months) it's much more reliable.

This commit adds support for multiple URLs in the manifest and the
lockfile. See examples/debian_snapshot/bullseye.yaml for a sample usage.